revert: move slash commands back into ChatScreen, remove mixin
The SlashCommandsMixin added complexity (type stubs, pragma comments,
split patch paths) without enough benefit. Commands back in chat.py.
